feat: rest area 템플릿 생성
