+++ src/pages/templates/rest-area-detail/index.ts
@@ -1,0 +1,1 @@
+export { RestAreaDetail } from './RestAreaDetail';
