+++ src/routers/index.tsx
@@ -8,7 +8,9 @@ import { Outlet, createBrowserRouter } from 'react-router-dom';
 
 import App from '#/App';
 import { LocationSearch } from '#/pages/location-search';
+import { RestAreaFuelPage } from '#/pages/rest-area-fuel';
 import { MobileView } from '#/pages/templates/mobile-view';
+import { RestAreaDetail } from '#/pages/templates/rest-area-detail';
 import { GlobalStyle } from '#/styles/global';
 import { theme } from '#/styles/theme';
 
@@ -53,6 +55,25 @@ export const applicationRouter: ReturnType<typeof createBrowserRouter> =
                     path: '/location-search',
                     errorElement: <div>에러</div>,
                     element: <LocationSearch />,
+                },
+                {
+                    path: '/rest-area/:restAreaId',
+                    errorElement: <div>에러</div>,
+                    element: <RestAreaDetail />,
+                    children: [
+                        {
+                            path: 'foods',
+                            element: <div>food</div>,
+                        },
+                        {
+                            path: 'fuel-parking',
+                            element: <RestAreaFuelPage />,
+                        },
+                        {
+                            path: 'other-information',
+                            element: <div>other information</div>,
+                        },
+                    ],
                 },
             ],
         },
